requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-binance==1.0.19
python-dotenv==0.19.0
pydantic==1.8.2
websockets==10.0
python-telegram-bot==20.6